Добавляет класс Smart
