+++ src/view/smart.js
@@ -1,0 +1,42 @@
+import Abstract from './abstract';
+
+export default class Smart extends Abstract {
+  constructor() {
+    super();
+    this._data = {};
+  }
+
+  updateData(update, justDataUpdating) {
+    if (!update) {
+      return;
+    }
+
+    this._data = Object.assign(
+        {},
+        this._data,
+        update
+    );
+
+    if (justDataUpdating) {
+      return;
+    }
+
+    this.updateElement();
+  }
+
+  updateElement() {
+    let prevElement = this.getElement();
+    const parent = prevElement.parentElement;
+    this.removeElement();
+
+    const newElement = this.getElement();
+
+    parent.replaceChild(newElement, prevElement);
+
+    this.restoreOn();
+  }
+
+  restoreOn() {
+    throw new Error(`Abstract method not implemented: restoreOn`);
+  }
+}
